updated tag styling in player
