+++ src/components/Player.tsx
@@ -16,6 +16,7 @@ import * as Slider from '@radix-ui/react-slider';
 import * as Tooltip from '@radix-ui/react-tooltip';
 import { $pagePlayersState } from '../store.ts';
 import { useStore } from '@nanostores/react';
+import { formatTimestamp } from '../utils/player.ts';
 
 interface Props {
   url: string;
@@ -25,23 +26,6 @@ interface Props {
   position?: number;
   id: string;
 }
-
-const formatTimestamp = (seconds: number, includeMs = true) => {
-  const date = new Date(seconds * 1000);
-  const hh = date.getUTCHours().toString().padStart(2, '0');
-  const mm = date.getUTCMinutes().toString().padStart(2, '0');
-  const ss = date.getUTCSeconds().toString().padStart(2, '0');
-  let ms;
-
-  let str = `${hh}:${mm.toString().padStart(2, '0')}:${ss}`;
-
-  if (includeMs) {
-    ms = date.getUTCMilliseconds().toString().padStart(3, '0');
-    str = `${str}:${ms}`;
-  }
-
-  return str;
-};
 
 const ReactPlayer = _ReactPlayer as unknown as React.FC<ReactPlayerProps>;
